Major update to DALECS_v2.py
This will likely be the last change to DALECS_v2.py before we replace
DALECS.py with it. The main changes are:

- integrate BS.py into DALECS_v2.py and use for all Biot-Savart calls
- add member functions to transform between Cartesian and spherical
  coordinates
- break dalecs into ionospheric, field-aligned, and equatorial currents,
  in addition to type1 and type2 loops

The last item in particular resulted in substantial slow-downs when
instantiating and processing DALECS objects (probably because of nested
loops over nested object arrays). This can almost certainly be sped
up, but for now, commit something that "just works".