Initial Biot-Savart as matrix multiplication
This is a major change that adds the capability to the DALECS class
to accelerate Biot-Savart calculations using a matrix multiplication
once the static transform matrix has been calculated. This was very
successful, and although generating the transform matrix is still
quite slow, once it exists, BS calculations are sped up by a factor
of 1000-2000x.

The problem is that in testing, it became apparent that there was
one major problem with deltaBTimeSeries.extractQuantities() that
prevented getting accurate results from this new method. For the
new method to work, the OBS locations could not change wrt the
DALECS. This was NOT the case when calling extractQuanties() with
GEO coordinates specified.

So, this commit will get everything saved that is working in DALECS,
plus the latest version of extractQuantities(), which does NOT give
the correct results if GEO coordinates are specified. After that, I
will embark on a truly major overhaul of the deltaBTimeSeries module,
and especially the extractQuantities() method.