Update deltaBTimeSeries for BS matrix
Previously, deltaBTimeSeries.py was not using DALECS' BS matrix
acceleration correctly, at least in GEO coordinates, because the
obs locations would change with each time step. This update is
a major overhaul in the logic of the extractQuantities() method
whereby all DALECS (and associated BS calculations) are created
in the coordinate frame specified by the user. This requires
interpolation of input currents, and thus some loss of accuracy,
but as long as the input ionospheric grids are of a reasonable
resolution, this is minimal. On the up-side, it is now possible
to request output on a geomagnetic grid, not just geographic and
solar magnetic.

This update also included several new keyword arguments to give
the user more control over which inputs are read in, which BS
integration method is used, and what obs grid coordinates are
used. I also cleaned up the command line interface a bit, and
fixed some small problems with related time series plot
generation.

The initial step of a loop that uses the BS matrix method is
still quite slow, but there is probably a fix to address much
of this slowness.