Optimized bs_cart(), cleaned up other stuff
There was a commented block of code in DALECS.py's bs_cart() method
that was a vectorized version of the looping BS integrator. It was
much more efficient, but at the cost of memory usage, making it
actually slower for large problems due to inefficiencies in memory
allocation, or on laptop, disk swapping. It finally occured to me
to create an if/then/else condition that checked the expected size
of the matrices to be created, and if they were over a threshold,
the method would switch to the looping method, and avoid the worst
of the memory issues (at the cost of a slow-but-not-as-slow process).
Also, I streamlined the BS wrappers inside the dalecs class so that
they didn't need to loop quite as many times.

Finally, I cleaned up deltaBTimeSeries.py just a bit, and I cleaned
up deltaB_IPyParallel.ipynb notebook considerably. I think this is
getting very close to ready to merge back in with the main LFM repo.