Initial attempt to Pool() with deltaBTimeSeries.py
This was an initial attemt to use multiprocessing.Pool() to
parallelize over virtual observatories. For 26 observatories
across 2 cpus, the performance dropped by a factor of 5. This
is almost certainly a knwn problem with Pool when distributing
across large objects like DALECS. Each complete object must
get serialized then copied to the worker processes, then the
output must be serialized and passed back. This may be less
of an issue for a larger number of observatories, where the
processing time will be more comparable to the time needed
to copy objects to worker processes.

There are other libraries that have similar functionality, but
which are more optimized for large data objects. Dask is one
such example. There is also mutli-threading, which would allow
objects to simply be shared without copying to worker threads.

Ultimately, I left the code in, but removed the ncpus option
and hardcoded ncpus to 1, just in case we want to return to
this sometime in the future.