Latest attempt to parallelize across processes
This is a little better than the previous attempt, since it copies
a DALECS to the global namespace of the worker processes just once,
then runs DALECS instance methods on those workers. However, it is
ugly, and probably not particularly stable. What's more, it still
does not allow the matrix optimization for DALECS to be used, so
even with perfectly linear cpu scaling, the improved performance
comes nowhere near the 1000x speedup obtained with the matrix
method.