Move obs-based parallelization into DALECS module
Multi-process parallelization in the deltaBTimeSeries.py module,
specifically inisde the `extractQuantities()` method, were ugly
and probably not particularly safe, since in order to get decent
performance, I had to forgoe the context manager, and manage all
Pools myself. The reason was that bandwidth requried to transfer
a DALECS object to the worker processes was immense, so we chose
to copy it once into the worker processes' global namespaces,
and modify it in place each time a new current distribution
became available. It scaled reasonably well, but it was not
going to be very interpretable or maintainable.

Here, I attempt to move the Pools into lower level methods of
the DALECS module rather than parallelize at the class instance
level like before. The methods and data passed to the worker
processes are now much smaller, requiring less bandwidth, and
can therefore be called using the context manager, which should
(hypothetically) take care of cleaning up after itself once a
Pool of processes is finished.

In addition, I use shared memory to hold the DALECS current
distributions becaue these will be identical for a single time
step, regardless of the observatory locations. This offers a
slight performance increase (10-20%), but more importantly, it
saves memory, which is critial for typical use-cases. It may
be necessary to do some follow-up memory profiling to see if
there are any memory leaks, since I only actually considered
this potential problem after I started typing up this rather
long commit message.